chore: move remote engine into engine page instead extension page
