+++ web/screens/Settings/Engines/index.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import React, { useCallback, useState } from 'react'
+import React, { useCallback, useEffect, useState } from 'react'
 
 import { InferenceEngine } from '@janhq/core'
 import { Button, ScrollArea, Input, Switch, Badge } from '@janhq/joi'
@@ -7,13 +7,28 @@ import { Button, ScrollArea, Input, Switch, Badge } from '@janhq/joi'
 import { useAtom, useSetAtom } from 'jotai'
 import { SearchIcon, SettingsIcon } from 'lucide-react'
 
+import { marked } from 'marked'
+
+import SetupRemoteModel from '@/containers/SetupRemoteModel'
+
 import {
   useGetDefaultEngineVariant,
   useGetEngines,
 } from '@/hooks/useEngineManagement'
 
-import { showSettingActiveLocalEngineAtom } from '@/helpers/atoms/Extension.atom'
+import { formatExtensionsName } from '@/utils/converter'
+
+import { extensionManager } from '@/extension'
+import Extension from '@/extension/Extension'
+import {
+  inActiveEngineProviderAtom,
+  showSettingActiveLocalEngineAtom,
+} from '@/helpers/atoms/Extension.atom'
 import { selectedSettingAtom } from '@/helpers/atoms/Setting.atom'
+
+type EngineExtension = {
+  provider: InferenceEngine
+} & Extension
 
 const EngineItems = ({ engine }: { engine: InferenceEngine }) => {
   const { defaultEngineVariant } = useGetDefaultEngineVariant(engine)
@@ -49,7 +64,7 @@ const EngineItems = ({ engine }: { engine: InferenceEngine }) => {
     [showSettingActiveLocalEngine, setShowSettingActiveLocalEngineAtom]
   )
   return (
-    <div className="flex w-full flex-col items-start justify-between py-3 sm:flex-row">
+    <div className="flex w-full flex-col items-start justify-between border-b border-[hsla(var(--app-border))] py-3 sm:flex-row">
       <div className="w-full flex-shrink-0 space-y-1.5">
         <div className="flex items-center justify-between gap-x-2">
           <div>
@@ -89,6 +104,63 @@ const EngineItems = ({ engine }: { engine: InferenceEngine }) => {
 const Engines = () => {
   const [searchText, setSearchText] = useState('')
   const { engines } = useGetEngines()
+  const [engineActiveExtensions, setEngineActiveExtensions] = useState<
+    EngineExtension[]
+  >([])
+  const [inActiveEngineProvider, setInActiveEngineProvider] = useAtom(
+    inActiveEngineProviderAtom
+  )
+
+  useEffect(() => {
+    const getAllSettings = async () => {
+      const extensionsMenu = []
+      const engineMenu = []
+      const extensions = extensionManager.getAll()
+
+      for (const extension of extensions) {
+        const settings = await extension.getSettings()
+        if (
+          typeof extension.getSettings === 'function' &&
+          'provider' in extension &&
+          typeof extension.provider === 'string'
+        ) {
+          if (
+            (settings && settings.length > 0) ||
+            (await extension.installationState()) !== 'NotRequired'
+          ) {
+            engineMenu.push({
+              ...extension,
+              provider:
+                'provider' in extension &&
+                typeof extension.provider === 'string'
+                  ? extension.provider
+                  : '',
+            })
+          }
+        } else {
+          extensionsMenu.push({
+            ...extension,
+          })
+        }
+      }
+
+      setEngineActiveExtensions(engineMenu as any)
+    }
+    getAllSettings()
+  }, [])
+
+  const onSwitchChange = useCallback(
+    (name: string) => {
+      if (inActiveEngineProvider.includes(name)) {
+        setInActiveEngineProvider(
+          [...inActiveEngineProvider].filter((x) => x !== name)
+        )
+      } else {
+        setInActiveEngineProvider([...inActiveEngineProvider, name])
+      }
+    },
+    [inActiveEngineProvider, setInActiveEngineProvider]
+  )
 
   return (
     <ScrollArea className="h-full w-full">
@@ -110,7 +182,7 @@ const Engines = () => {
       </div>
 
       <div className="block w-full px-4">
-        <div className="mb-3 mt-4 border-b border-[hsla(var(--app-border))] pb-4">
+        <div className="mb-3 mt-4 pb-4">
           <h6 className="text-xs text-[hsla(var(--text-secondary))]">
             Local Engine
           </h6>
@@ -123,6 +195,65 @@ const Engines = () => {
               })}
         </div>
       </div>
+
+      {engineActiveExtensions.length !== 0 && (
+        <div className="mt-4 block w-full px-4">
+          <div className="mb-3 mt-4 pb-4">
+            <h6 className="text-xs text-[hsla(var(--text-secondary))]">
+              Remote Engine
+            </h6>
+            {engineActiveExtensions
+              .filter((x) => x.name.includes(searchText.toLowerCase().trim()))
+              .sort((a, b) => a.provider.localeCompare(b.provider))
+              .map((item, i) => {
+                return (
+                  <div
+                    key={i}
+                    className="flex w-full flex-col items-start justify-between border-b border-[hsla(var(--app-border))] py-3 sm:flex-row"
+                  >
+                    <div className="w-full flex-shrink-0 space-y-1.5">
+                      <div className="flex items-center justify-between gap-x-2">
+                        <div className="flex items-center gap-x-2">
+                          <h6 className="line-clamp-1 font-semibold">
+                            {item.productName?.replace(
+                              'Inference Engine',
+                              ''
+                            ) ?? formatExtensionsName(item.name)}
+                          </h6>
+                          <Badge variant="outline" theme="secondary">
+                            v{item.version}
+                          </Badge>
+                          <p>{item.provider}</p>
+                        </div>
+                        <div className="flex items-center gap-x-2">
+                          <Switch
+                            checked={
+                              !inActiveEngineProvider.includes(item.provider)
+                            }
+                            onChange={() => onSwitchChange(item.provider)}
+                          />
+                          {!inActiveEngineProvider.includes(item.provider) && (
+                            <SetupRemoteModel engine={item.provider} />
+                          )}
+                        </div>
+                      </div>
+                      {
+                        <div
+                          className="w-full font-medium leading-relaxed text-[hsla(var(--text-secondary))] sm:w-4/5"
+                          dangerouslySetInnerHTML={{
+                            __html: marked.parse(item.description ?? '', {
+                              async: false,
+                            }),
+                          }}
+                        />
+                      }
+                    </div>
+                  </div>
+                )
+              })}
+          </div>
+        </div>
+      )}
     </ScrollArea>
   )
 }
