+++ web/screens/Settings/SettingLeftPanel/index.tsx
@@ -123,7 +123,7 @@ const SettingLeftPanel = () => {
         ).length > 0 && (
           <div className="mb-1 mt-4 px-2">
             <label className="text-xs font-medium text-[hsla(var(--text-secondary))]">
-              Model Providers
+              Remote Engine
             </label>
           </div>
         )}
